requirements/prod.txt
pandas
mecab-python
sentencepiece
tensor2tensor
tensorflow
tiny_tokenizer[all]